+++ client/src/components/Home/cardhome.module.css
@@ -1,0 +1,50 @@
+.card {
+  display: flex;
+  justify-content: center;
+  margin: 1rem;
+  background: #f8f8f8;
+  height: fit-content;
+  width: 280px;
+  /* box-shadow: -2px 10px 14px 0px #626262; */
+  border-radius: 8px;
+  transition: 0.5s;
+  cursor: pointer;
+  border: #dadada 1px solid;
+}
+
+.cards {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 32px;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+
+.card:hover {
+  box-shadow: -2px 10px 14px 0px #626262;
+
+  cursor: pointer;
+}
+
+.img {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 1rem;
+  width: 230px;
+  height: 230px;
+  object-fit: cover;
+  align-content: center;
+  border-radius: 9px;
+  border: 3px solid #dbdbdb;
+  /* border: 3px solid #2AADF6; /
+      / border: 3px solid #e56ad5 */
+}
+
+.name {
+  display: flex;
+  justify-content: center;
+}
